improve: Relocate MockedTransactionClient
Will be used for testing both the TransactionClient and the
MultiCallerClient.
