+++ test/mocks/index.ts
@@ -3,4 +3,5 @@ export * from "./MockHubPoolClient";
 export * from "./MockProfitClient";
 export * from "./MockAdapterManager";
 export * from "./MockTokenClient";
+export * from "./MockTransactionClient";
 export * from "./MockInventoryClient";
